Added multiselection.
Singular deselection from a multiselection is missing.

How it works:
Hold the SHIFT key and start clicking on the nodes, it will select all of them. Once you release the key, it will allow you to select any of the nodes to deselect the others, otherwise just press an empty space to deselect all of them.